src/mad_twinnet_train/requirements.txt
numpy>=1.16.3
torch>=1.1.0
scipy>=1.2.1
mir-eval>=0.4
pyyaml>=3.13
